encompass new feature in util function
