+++ orchestration/dags/jobs/dbt/dbt_dag_run.py
@@ -6,12 +6,11 @@ from common.config import (
     PATH_TO_DBT_PROJECT,
     PATH_TO_DBT_TARGET,
 )
-from common.dbt.utils import load_manifest
+from common.dbt.utils import create_nested_folder_groups, load_manifest
 from common.utils import get_airflow_schedule, waiting_operator
 
 from airflow import DAG
 from airflow.models import Param
-from airflow.models.baseoperator import chain
 from airflow.operators.bash_operator import BashOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.utils.dates import datetime, timedelta
@@ -24,6 +23,7 @@ default_args = {
     "project_id": GCP_PROJECT_ID,
     "on_failure_callback": task_fail_slack_alert,
 }
+
 
 dag = DAG(
     "dbt_run_dag",
@@ -253,70 +253,10 @@ with TaskGroup(group_id="snapshots", dag=dag) as snapshot_group:
         )
 
 # Trigger folder of models manually
-nested_folders = {
-    model_node: manifest["nodes"][model_node]["fqn"][1:-1] for model_node in dbt_models
-}
-task_groups = {}
 
 with TaskGroup(group_id="folders_manual_trigger", dag=dag) as trigger_block:
+    create_nested_folder_groups(dbt_models, model_op_dict, manifest, dag)
 
-    def create_nested_task_group(
-        folder_hierarchy, original_hierarchy, parent_group=None, current_path=""
-    ):
-        """
-        Recursively create nested TaskGroups based on folder hierarchy.
-        TODO: remove leaf TaskGroup and put the trigger operator in parent TaskGroup
-            hint: the recursive function is nihilpotent over original_hierarchy
-        """
-        if not folder_hierarchy:
-            return None, original_hierarchy
-
-        current_folder = folder_hierarchy[0]
-
-        # Build the full path for the group_id
-        current_path = (
-            f"{current_path}_{current_folder}" if current_path else current_folder
-        )
-        group_id = current_path
-
-        # Check if the group already exists
-        if group_id in task_groups:
-            tg, dummy_task = task_groups[group_id]
-        else:
-            # Create task group & corresponding trigger opperator
-            tg = TaskGroup(group_id=group_id, parent_group=parent_group, dag=dag)
-            dummy_task = DummyOperator(
-                task_id=f"trigger_{group_id}_folder", task_group=tg, dag=dag
-            )
-            # Add them to dictionnary
-            task_groups[group_id] = (tg, dummy_task)
-
-        # Recurse to create the nested TaskGroups
-        if len(folder_hierarchy) > 1:
-            _, original_hierarchy = create_nested_task_group(
-                folder_hierarchy[1:],
-                original_hierarchy,
-                parent_group=tg,
-                current_path=current_path,
-            )
-
-        return tg, original_hierarchy
-
-    # Create task groups for each folder hierarchy in nested_folders
-    for folder_hierarchy in nested_folders.values():
-        original_hierarchy = folder_hierarchy.copy()
-        create_nested_task_group(folder_hierarchy, original_hierarchy)
-
-    # Set the upstream for each task in model_op_dict to the corresponding trigger operator
-    for model_node, folder_hierarchy in nested_folders.items():
-        # Chain trigger tasks of folder hierarchy
-        task_chain = [
-            task_groups["_".join(folder_hierarchy[: i + 1])][1]
-            for i, _ in enumerate(folder_hierarchy)
-        ]
-        chain(*task_chain)
-        # Set model execution task's upstream to leaf folder trigger task
-        task_chain[-1] >> model_op_dict[model_node]
 
 (
     start
